decouple booting wpds in dev by default
