+++ package.json
@@ -15,6 +15,7 @@
     "babel-core": "^5.6.20",
     "bluebird": "^2.9.34",
     "bootstrap": "twbs/bootstrap-sass#v3.3.5",
+    "chalk": "^1.1.1",
     "coinstac-distributed-algorithm-set": "git+https://github.com/MRN-Code/coinstac-distributed-algorithm-set.git",
     "config": "^1.14.0",
     "errio": "^1.2.1",
@@ -81,7 +82,7 @@
     "lint": "echo 'run linter!'",
     "validate": "echo 'run validator!'",
     "test": "grunt test",
-    "webpack": "node webpack-server.js",
+    "webpack": "node webpack-server.js --development",
     "debug": "electron --debug-brk=5858 . --development",
     "start": "npm run shownotes && electron . --development",
     "shownotes": "node ./dev-notes.js",
